rename compat to version
